refactor(Game): 패키지 구조 및 코드 리팩토링
